Update config to match 3.0 content (disables cross-faction mail, characters until 3.2). Fix launch warnings because of type mismatch.

+++ src/server/game/Skills/SkillExtraItems.cpp
@@ -182,7 +182,7 @@ void LoadSkillExtraItemTable()
             continue;
         }
 
-        uint8 additionalMaxNum = fields[3].GetInt32();
+        uint8 additionalMaxNum = fields[3].GetInt8();
         if (!additionalMaxNum)
         {
             sLog->outError("Skill specialization %u has 0 max number of extra items in `skill_extra_item_template`!", spellId);
